% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_index_etc.R
\name{calc_index_etc}
\alias{calc_index_etc}
\alias{calc_index_etc.WithParent}
\alias{calc_index_etc.WithoutParent}
\title{Calculate Smoothed SNP Index and Euclidean Distance Using Sliding Window}
\usage{
calc_index_etc(x, ...)

\method{calc_index_etc}{WithParent}(x, outPrefix, winSize, winStep)

\method{calc_index_etc}{WithoutParent}(x, outPrefix, winSize, winStep)
}
\arguments{
\item{x}{The QTLseq S3 object.}

\item{outPrefix}{The prefix of <outPrefix>.SlidingWindow.txt and <outPrefix>.SlidingWindow.csv}

\item{winSize}{Window size for sliding window.}

\item{winStep}{Window step for sliding window.}
}
\value{
A QTLseq S3 object with sliding window of SNP index and Euclidean Distance.
}
\description{
Calculate Smoothed SNP Index and Euclidean Distance Using Sliding Window
}
\examples{
NULL
}
