% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_dp.R
\name{export_dp}
\alias{export_dp}
\alias{export_dp.WithParent}
\alias{export_dp.WithoutParent}
\title{Export Allele Depth}
\usage{
export_dp(x, ...)

\method{export_dp}{WithParent}(x, outPrefix)

\method{export_dp}{WithoutParent}(x, outPrefix)
}
\arguments{
\item{x}{A QTLseq S3 object.}

\item{outPrefix}{The prefix of output file.}
}
\description{
export_dp writes allele depth information to <outPrefix>.Depth_information.txt and <outPrefix>.Depth_information.csv
}
\examples{
library(easyQTLseq)
# Example with sample data from a GATK table.
file_path <- system.file("extdata", "subset.table.gz", package = "easyQTLseq")
# readr::read_tsv() has a faster speed than read.table() when reading a file.
data <- readr::read_tsv(file = file_path)
x <- select_sample_and_SNP(data = data, highP = "qY", lowP = "R3", highB = "Y", lowB = "R", popType = "F2", bulkSize = c(30, 30))
x_filter <- filterDP(x = x)
export_dp(x = x, outPrefix = "outfile")
}
