% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_figure.R
\name{export_figure}
\alias{export_figure}
\alias{export_figure.WithParent}
\alias{export_figure.WithoutParent}
\title{Export Figures of Sliding Window SNP Index and Euclidean Distance}
\usage{
export_figure(x, ...)

\method{export_figure}{WithParent}(
  x,
  outPrefix,
  targetChr,
  chrLabel,
  minN,
  width,
  height,
  color = c("#4197d8", "#f8c120", "#413496", "#495226", "#d60b6f", "#e66519", "#d581b7",
    "#83d3ad", "#7c162c", "#26755d")
)

\method{export_figure}{WithoutParent}(
  x,
  outPrefix,
  targetChr,
  chrLabel,
  minN,
  width,
  height,
  color = c("#4197d8", "#f8c120", "#413496", "#495226", "#d60b6f", "#e66519", "#d581b7",
    "#83d3ad", "#7c162c", "#26755d")
)
}
\arguments{
\item{x}{A QTLseq S3 object with sliding window of SNP index and Euclidean Distance.}

\item{outPrefix}{The prefix of figures.}

\item{targetChr}{Target chromosome to be drawn in figures, default is all chromosomes in the data.}

\item{chrLabel}{The label for chromosome shown in figures, default is chromosome names in the data.}

\item{minN}{The windows containing SNPs less than minN will be omitted in figures.}

\item{width}{The width of figures.}

\item{height}{The height of figures.}

\item{color}{A vector of colors for chromosomes.}
}
\description{
Export Figures of Sliding Window SNP Index and Euclidean Distance
}
\examples{
NULL
}
