% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQTL_and_exportFigure.R
\name{getQTL_and_exportFigure}
\alias{getQTL_and_exportFigure}
\alias{getQTL_and_exportFigure.WithParent}
\title{Get QTL Exceed 95\% and 99\% Confidence Interval}
\usage{
getQTL_and_exportFigure(x, ...)

\method{getQTL_and_exportFigure}{WithParent}(x, outPrefix, minN)
}
\arguments{
\item{x}{The QTLseq S3 object.}

\item{outPrefix}{The prefix of outfiles.}

\item{minN}{The minimum number of SNP in a window.}
}
\description{
This function will export <outPrefix>.95CI.csv, <outPrefix>.99CI.csv and figures of target chromosomes.
}
\examples{
#' library(easyQTLseq)
# Example with sample data from a GATK table.
file_path <- system.file("extdata", "subset.table.gz", package = "easyQTLseq")
# readr::read_tsv() has a faster speed than read.table() when reading a file.
data <- readr::read_tsv(file = file_path)
x <- select_sample_and_SNP(data = data, highP = "qY", lowP = "R3", highB = "Y", lowB = "R", popType = "F2", bulkSize = c(30, 30))
x_filter <- filterDP(x = x)
x_filter <- calc_index_etc(x = x_filter, outPrefix = "outprefix", winSize = 2000000, winStep = 20000)
getQTL_and_exportFigure(x = x_filter, outPrefix = "outprefix", minN = 20)
}
