% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_sample_and_SNP.R
\name{select_sample_and_SNP}
\alias{select_sample_and_SNP}
\title{Select bulk samples and parent samples}
\usage{
select_sample_and_SNP(
  data,
  highP,
  lowP,
  highB,
  lowB,
  popType,
  bulkSize,
  minGQ = 10,
  refFreq = 0.3,
  chrLen
)
}
\arguments{
\item{data}{A data set generated by GATK's `VariantsToTable` module.}

\item{highP}{The name of high phenotype parent.}

\item{lowP}{The name of low phenotype parent.}

\item{highB}{The name of high phenotype bulk.}

\item{lowB}{The name of low phenotype bulk.}

\item{popType}{Population type, RIL or F2.}

\item{bulkSize}{A integer vector of two element, represent high bulk size and low bulk size.}

\item{minGQ}{Minimum GQ for parent.}

\item{refFreq}{SNP will be discarded if ERF Allelic Depth ratio is less than `refFreq` or more than `1-refFreq` in both bulks.}
}
\value{
A table SNP position, allelic depth.
}
\description{
`select_sample_and_SNP()` select `highParent.GT`, `highParent.AD`, `highParent.GQ`, `lowParent.GT`, `lowParent.AD`, `lowParent.GQ`,
`highBulk.GT`, `highBulk.AD`, `highBulk.GQ`, `lowBulk.GT`, `lowBulk.AD` and `lowBulk.GQ` from a dataset generated from GATK' `VariantsToTable` module.
Effective SNP sites are selected according both parents and bulks information. Then bulk AD will be reassignment according parent genotype.
}
\examples{
library(easyQTLseq)
# Example with sample data from a GATK table.
file_path <- system.file("extdata", "subset.table.gz", package = "easyQTLseq")
# readr::read_tsv() has a faster speed than read.table() when reading a file.
data <- readr::read_tsv(file = file_path)
x <- importData(data = data, highP = "qY", lowP = "R3", highB = "Y", lowB = "R", popType = "F2", bulkSize = c(30, 30))
}
