% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNP_distribution.R
\name{SNP_distribution}
\alias{SNP_distribution}
\alias{SNP_distribution.QTLseq}
\title{Analyzing the Distribution of SNPs on Chromosomes}
\usage{
SNP_distribution(x, ...)

\method{SNP_distribution}{QTLseq}(x, outPrefix, targetChr, chrLabel)
}
\arguments{
\item{x}{The QTLseq S3 object.}

\item{outPrefix}{The prefix of output file. <outPrefix>.SNP_number_per_chr.txt|csv and <outPrefix>.SNP_distribution_histogram.pdf|png}

\item{targetChr}{Only chromosomes in targetChr will be analyzed, default is all chromosomes.}

\item{chrLabel}{The labels of chromosomes, default is the same of targetChr.}
}
\description{
The SNP number in every 1 Mb bins will be shown in histogram along chromosome.
}
\examples{
library(easyQTLseq)
# Example with sample data from a GATK table.
file_path <- system.file("extdata", "subset.table.gz", package = "easyQTLseq")
# readr::read_tsv() has a faster speed than read.table() when reading a file.
data <- readr::read_tsv(file = file_path)
x <- select_sample_and_SNP(data = data, highP = "qY", lowP = "R3", highB = "Y", lowB = "R", popType = "F2", bulkSize = c(30, 30))
x_filter <- filterDP(x = x)
SNP_distribution(x = x_filter, outPrefix = "outprefix",
                 targetChr = c("scaffoldA01", "scaffoldA07", "scaffoldA09"),
                 chrLabel = c("A01", "A07", "A09"))
}
