% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2table.R
\name{vcf2table}
\alias{vcf2table}
\title{Convert vcfR Object to A Table}
\usage{
vcf2table(x)
}
\arguments{
\item{x}{A vcfR object}
}
\value{
A table
}
\description{
vcf2table is used to convert a vcfR object to a table, then this table can be used for select_sample_and_SNP() to construct QTLseq object.
}
\examples{
library(vcfR)
library(easyQTLseq)
file_path <- system.file("extdata", "A07.SNPs.vcf.gz", package = "easyQTLseq")
x <- read.vcfR(file = file_path)
data <- vcf2table(x = x)
}
