% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_index_etc.R
\name{calc_index_etc}
\alias{calc_index_etc}
\alias{calc_index_etc.WithParent}
\alias{calc_index_etc.WithoutParent}
\title{Calculate Smoothed SNP Index and Euclidean Distance Using Sliding Window}
\usage{
calc_index_etc(x, ...)

\method{calc_index_etc}{WithParent}(x, outPrefix, winSize, winStep)

\method{calc_index_etc}{WithoutParent}(x, outPrefix, winSize, winStep)
}
\arguments{
\item{x}{The QTLseq S3 object.}

\item{outPrefix}{The prefix of <outPrefix>.SlidingWindow.txt and <outPrefix>.SlidingWindow.csv}

\item{winSize}{Window size for sliding window.}

\item{winStep}{Window step for sliding window.}
}
\value{
A QTLseq S3 object with sliding window of SNP index and Euclidean Distance.
}
\description{
Calculate Smoothed SNP Index and Euclidean Distance Using Sliding Window
}
\examples{
library(easyQTLseq)
# Example with sample data from a GATK table.
file_path <- system.file("extdata", "subset.table.gz", package = "easyQTLseq")
# readr::read_tsv() has a faster speed than read.table() when reading a file.
data <- readr::read_tsv(file = file_path)
x <- select_sample_and_SNP(data = data, highP = "qY", lowP = "R3", highB = "Y", lowB = "R", popType = "F2", bulkSize = c(30, 30))
x_filter <- filterDP(x = x)
x_filter <- calc_index_etc(x = x_filter, outPrefix = "outprefix", winSize = 2000000, winStep = 200000)
}
